backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.20
pydantic==2.9.2
groq>=0.12.0
python-dotenv==1.0.0
httpx>=0.27.2
